+++ components/brave_ads/core/internal/resources/behavioral/purchase_intent/purchase_intent_info.h
@@ -6,7 +6,6 @@
 #ifndef BRAVE_COMPONENTS_BRAVE_ADS_CORE_INTERNAL_RESOURCES_BEHAVIORAL_PURCHASE_INTENT_PURCHASE_INTENT_INFO_H_
 #define BRAVE_COMPONENTS_BRAVE_ADS_CORE_INTERNAL_RESOURCES_BEHAVIORAL_PURCHASE_INTENT_PURCHASE_INTENT_INFO_H_
 
-#include <cstdint>
 #include <string>
 #include <vector>
 
@@ -32,7 +31,7 @@ struct PurchaseIntentInfo final {
   static base::expected<PurchaseIntentInfo, std::string> CreateFromValue(
       base::Value::Dict dict);
 
-  uint16_t version = 0;
+  int version = 0;
   std::vector<PurchaseIntentSiteInfo> sites;
   std::vector<PurchaseIntentSegmentKeywordInfo> segment_keywords;
   std::vector<PurchaseIntentFunnelKeywordInfo> funnel_keywords;
