+++ components/brave_ads/core/internal/user_attention/user_activity/user_activity_util.h
@@ -6,7 +6,6 @@
 #ifndef BRAVE_COMPONENTS_BRAVE_ADS_CORE_INTERNAL_USER_ATTENTION_USER_ACTIVITY_USER_ACTIVITY_UTIL_H_
 #define BRAVE_COMPONENTS_BRAVE_ADS_CORE_INTERNAL_USER_ATTENTION_USER_ACTIVITY_USER_ACTIVITY_UTIL_H_
 
-#include <cstdint>
 #include <string>
 
 #include "brave/components/brave_ads/core/internal/user_attention/user_activity/user_activity_event_info.h"
@@ -15,10 +14,10 @@
 
 namespace brave_ads {
 
-int GetNumberOfTabsOpened(const UserActivityEventList& events);
+size_t GetNumberOfTabsOpened(const UserActivityEventList& events);
 
-int GetNumberOfUserActivityEvents(const UserActivityEventList& events,
-                                  UserActivityEventType event_type);
+size_t GetNumberOfUserActivityEvents(const UserActivityEventList& events,
+                                     UserActivityEventType event_type);
 base::TimeDelta GetTimeSinceLastUserActivityEvent(
     const UserActivityEventList& events,
     UserActivityEventType event_type);
